Feature(src/modules/modalities/useCases/deleteModality) Create deleteModality UseCase
Is possibile to delete one modality

PLAYT39

+++ src/modules/modalities/useCases/deleteModality/DeleteModalityUseCase.ts
@@ -5,11 +5,17 @@ import { ModalitiesRespository } from "../../repositories/implementations/Modali
 @injectable()
 class DeleteModalityUseCase {
   constructor(
-    @inject("ModalityRepositories")
+    @inject("ModalitiesRespository")
     private modalityRepository: ModalitiesRespository
   ) {}
 
   async execute({ id }): Promise<void> {
+    const modalityAlredyExists = await this.modalityRepository.findById(id);
+
+    if (!modalityAlredyExists) {
+      throw new Error("Esta modalidade não existe e não pode ser deletada.");
+    }
+
     await this.modalityRepository.delete(id);
   }
 }
